% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultOrd-package.R
\docType{package}
\name{MultOrd-package}
\alias{MultOrd-package}
\title{Model Multivariate Ordinal Responses Including Response Styles}
\description{
A model for multivariate ordinal responses. The response is modelled 
using a mixed model approach that is also capable of the inclusion 
of response style effects of the respondents.
}
\examples{
\dontrun{
data(tenseness)

## create a small subset of the data to speed up calculations
set.seed(1860)
tenseness <- tenseness[sample(1:nrow(tenseness), 300),]

## scale all metric variables to get comparable parameter estimates
tenseness$Age <- scale(tenseness$Age)
tenseness$Income <- scale(tenseness$Income)

## two formulas, one without and one with explanatory variables (gender and age)
f.tense0 <- as.formula(paste("cbind(",paste(names(tenseness)[1:4],collapse=","),") ~ 1"))
f.tense1 <- as.formula(paste("cbind(",paste(names(tenseness)[1:4],collapse=","),") ~ Gender + Age"))



####
## Adjacent Categories Models
####

## Multivariate adjacent categories model, without response style, without explanatory variables
m.tense0 <- multord(f.tense0, data = tenseness, control = ctrl.multord(RS = FALSE))
m.tense0

## Multivariate adjacent categories model, with response style as a random effect, without explanatory variables
m.tense1 <- multord(f.tense0, data = tenseness)
m.tense1

## Multivariate adjacent categories model, with response style as a random effect, 
## without explanatory variables for response style BUT for location
m.tense2 <- multord(f.tense1, data = tenseness, control = ctrl.multord(XforRS = FALSE))
m.tense2

## Multivariate adjacent categories model, with response style as a random effect, with explanatory variables for location AND response style
m.tense3 <- multord(f.tense1, data = tenseness)
m.tense3

plot(m.tense3)



####
## Cumulative Models
####

## Multivariate cumulative model, without response style, without explanatory variables
m.tense0.cumul <- multord(f.tense0, data = tenseness, control = ctrl.multord(RS = FALSE), model = "cumulative")
m.tense0.cumul

## Multivariate cumulative model, with response style as a random effect, without explanatory variables
m.tense1.cumul <- multord(f.tense0, data = tenseness, model = "cumulative")
m.tense1.cumul

## Multivariate cumulative model, with response style as a random effect, 
## without explanatory variables for response style BUT for location
m.tense2.cumul <- multord(f.tense1, data = tenseness, control = ctrl.multord(XforRS = FALSE), model = "cumulative")
m.tense2.cumul

## Multivariate cumulative model, with response style as a random effect, with explanatory variables for location AND response style
m.tense3.cumul <- multord(f.tense1, data = tenseness, model = "cumulative")
m.tense3.cumul

plot(m.tense3.cumul)
}
}
\seealso{
\code{\link{multord}}  \code{\link{ctrl.multord}}  \code{\link{plot.MultOrd}}
}
\author{
Gunther Schauberger\cr \email{gunther.schauberger@tum.de}\cr
\url{https://www.researchgate.net/profile/Gunther_Schauberger2}
}
\keyword{adjacent}
\keyword{categories}
\keyword{cumulative}
\keyword{multivariate}
\keyword{ordinal}
\keyword{response}
\keyword{style}
